Moved HostingPlatform model to separate app
