+++ FOSSDB_web/apps/fossdb/admin.py
@@ -1,12 +1,8 @@
 from django.contrib import admin
+from hosting_platform.admin import ProjectHostingPlatformInline
 from programming_language.admin import ProjectProgrammingLanguageInline
 
-from .models import HostingPlatform, License, Project, ProjectHostingPlatform
-
-
-class ProjectHostingPlatformInline(admin.TabularInline):
-    model = ProjectHostingPlatform
-    extra = 1
+from .models import Project
 
 
 class ProjectAdmin(admin.ModelAdmin):
@@ -20,7 +16,5 @@ class ProjectAdmin(admin.ModelAdmin):
         return " | ".join([i.hosting_platform.hosting_platform for i in object.projecthostingplatform_set.all()])
 
 
-admin.site.register(ProgrammingLanguage)
-admin.site.register(HostingPlatform)
 admin.site.register(Project, ProjectAdmin)
 admin.site.register(Tag)
