+++ FOSSDB_web/apps/hosting_platform/models.py
@@ -1,3 +1,17 @@
 from django.db import models
 
-# Create your models here.
+
+class HostingPlatform(models.Model):
+    hosting_platform = models.CharField(max_length=100)
+
+    def __str__(self):
+        return self.hosting_platform
+
+
+class ProjectHostingPlatform(models.Model):
+    project = models.ForeignKey("Project", on_delete=models.CASCADE)
+    hosting_platform = models.ForeignKey(HostingPlatform, on_delete=models.CASCADE)
+    url = models.URLField()
+
+    def __str__(self):
+        return f"{self.project} | {self.hosting_platform}"
